fix: resolve lint issues — unused params, exhaustive switches, package comments, and more
- unused-parameter (revive): rename 327 unused params to _ prefix
- package-comments (revive/ST1000): add package doc comments to 24 packages
- exhaustive: add missing switch cases for StateOpen, JitterNone, ReplayViewDetail, KeyType
- unconvert: remove 6 unnecessary type conversions
- ineffassign: fix 7 ineffectual assignments in tests
- wastedassign: fix redundant prefix assignment
- unparam: remove/suppress 29 unused function parameters
- Fix test struct literal ordering after fieldalignment reorder

+++ internal/storage/disk_unix.go
@@ -34,9 +34,9 @@ func getDiskUsagePlatform(path string) (total, used, available uint64, err error
 	// Bfree: Free blocks
 	// Bavail: Available blocks (for non-root)
 	blockSize := uint64(stat.Bsize)
-	total = blockSize * uint64(stat.Blocks)
-	freeBytes := blockSize * uint64(stat.Bfree)
-	available = blockSize * uint64(stat.Bavail)
+	total = blockSize * stat.Blocks
+	freeBytes := blockSize * stat.Bfree
+	available = blockSize * stat.Bavail
 	used = total - freeBytes
 
 	return total, used, available, nil
